Validate when baker or delegation amount is in overflow
