+++ app/src/main/java/com/concordium/wallet/ui/bakerdelegation/common/BaseDelegationBakerRegisterAmountActivity.kt
@@ -1,8 +1,10 @@
 package com.concordium.wallet.ui.bakerdelegation.common
 
 import com.concordium.wallet.R
+import com.concordium.wallet.data.util.CurrencyUtil
 import com.concordium.wallet.uicore.view.SegmentedControlView
-import kotlinx.android.synthetic.main.activity_baker_registration_amount.restake_options
+import kotlinx.android.synthetic.main.activity_baker_registration_amount.*
+import java.text.DecimalFormatSymbols
 
 abstract class BaseDelegationBakerRegisterAmountActivity(layout: Int, titleId: Int = R.string.app_name) :
     BaseDelegationBakerActivity(layout, titleId) {
@@ -40,5 +42,44 @@ abstract class BaseDelegationBakerRegisterAmountActivity(layout: Int, titleId: I
         return amountToStake > (viewModel.bakerDelegationData.account?.finalizedBalance ?: 0) * 0.95
     }
 
+    protected fun validateAmountInput(text: CharSequence?) {
+        if (text != null && (text.toString() == DecimalFormatSymbols.getInstance().decimalSeparator.toString() || text.filter { it == DecimalFormatSymbols.getInstance().decimalSeparator }.length > 1)) {
+            amount.setText(text.dropLast(1))
+        }
+        if (amount.text.isNotEmpty() && !amount.text.startsWith("Ͼ")) {
+            amount.setText("Ͼ".plus(amount.text.toString()))
+            amount.setSelection(amount.text.length)
+        }
+        setAmountHint()
+        if (amount.text.toString().isNotBlank() && amount.text.toString() != "Ͼ") {
+            val stakeAmountInputValidator = getStakeAmountInputValidator()
+            val stakeError = stakeAmountInputValidator.validate(CurrencyUtil.toGTUValue(amount.text.toString())?.toString(), fee)
+            if (stakeError != StakeAmountInputValidator.StakeError.OK) {
+                amount_error.text = stakeAmountInputValidator.getErrorText(this, stakeError)
+                showError(stakeError)
+            } else {
+                hideError()
+                loadTransactionFee()
+            }
+        } else {
+            hideError()
+        }
+    }
+
+    protected fun setAmountHint() {
+        when {
+            amount.text.isNotEmpty() -> {
+                amount.hint = ""
+            }
+            else -> {
+                amount.hint = "Ͼ0" + DecimalFormatSymbols.getInstance().decimalSeparator + "00"
+            }
+        }
+    }
+
     abstract fun getStakeAmountInputValidator(): StakeAmountInputValidator
+
+    abstract fun showError(stakeError: StakeAmountInputValidator.StakeError?)
+    abstract fun hideError()
+    abstract fun loadTransactionFee()
 }
